Add start up as spans
